Introduced a new scope - namely the class diagram scope - between artifact scope and the class symbols. This suits both the domain concept and our wording. Consequently, the package names of every symbol in the class diagram and the class diagram itself are the same. For example:
package p.q;
classdiagramm Uni {
    class Person;
    class Equipment;
}

The package name is p.q for all: Person, Equipment, Uni...The artifact scope contains the whole (file/model) content, i.e., the imports and the class diagram definition. Since we do not define a symbol for class diagrams, the lone class diagram SCOPE is defined in the artifact scope. This class diagram scope defines a new namespace - since it has a name  ("Uni"). In this scope all the class symbols are defined. The full name of those symbols is p.q.Uni.<ClassName>. Note that Uni is part of the full name, but not part of the package name. This is analogous to the relationship of classes and methods in Java. A method is in the same package as its defining class and its full name contains the name of the class, i.e., package.Class.method,

git-svn-id: https://sselab.de/lab2/private/svn/MontiCore/trunk/mclang/cd4analysis@21394 3ddcf219-3c89-4d16-bbf6-790a9aa18b9b